uint8 as default output dtype

+++ teigen/gui.py
@@ -666,8 +666,8 @@ class Teigen():
         logger.debug("before volume generate " + str(t1-t0))
         # postprocessing
         if "generate_volume" in dir(self.gen):
-            self.data3d = self.gen.generate_volume()
-            # self.data3d = self.gen.generate_volume(dtype="uint8")
+            # self.data3d = self.gen.generate_volume()
+            self.data3d = self.gen.generate_volume(dtype="uint8")
             self.voxelsize_mm = self.gen.voxelsize_mm
             postprocessing_params = self.config["postprocessing"]
             data3d = self.postprocessing(**postprocessing_params)
